Site Migration Update
Preparation for the eventual move from AWS to Github pages. It was nice to have, but now it's just easier to make a set it and forget it github page.

Having to handle and renew SSL certificates was always a pain.

Basically replicated what the original site had while trimming all the fat. I used to have an about me along with some references to some git repos. Really if people care they'll look. LinkedIn and Github have enough about me.